+++ assets/js/dashboard/query.js
@@ -29,6 +29,10 @@ export function parseQuery(querystring, site) {
       'utm_source': q.get('utm_source'),
       'utm_campaign': q.get('utm_campaign'),
       'referrer': q.get('referrer'),
+      'screen': q.get('screen'),
+      'browser': q.get('browser'),
+      'os': q.get('os'),
+      'country': q.get('country'),
       'page': q.get('page')
     }
   }